better error handling for tokens
