+++ cogs5e/sheetManager.py
@@ -23,8 +23,8 @@ from cogs5e.sheets.dicecloud import DicecloudParser
 from cogs5e.sheets.gsheet import GoogleSheet, extract_gsheet_id_from_url
 from utils.argparser import argparse
 from utils.constants import SKILL_NAMES
-from utils.functions import auth_and_chan, confirm, generate_token, get_positivity, list_get, search_and_select, \
-    try_delete
+from utils.functions import auth_and_chan, confirm, get_positivity, list_get, search_and_select, try_delete
+from utils.img import generate_token
 from utils.user_settings import CSetting
 
 log = logging.getLogger(__name__)
